+++ magenta/common/tf_utils.py
@@ -60,8 +60,8 @@ def log_loss(labels, predictions, epsilon=1e-7, scope=None, weights=None):
     ValueError: If the shape of `predictions` doesn't match that of `labels`.
   """
   with tf.name_scope(scope, "log_loss", (predictions, labels)):
-    predictions = tf.to_float(predictions)
-    labels = tf.to_float(labels)
+    predictions = tf.cast(predictions, tf.float32)
+    labels = tf.cast(labels, tf.float32)
     predictions.get_shape().assert_is_compatible_with(labels.get_shape())
     losses = -tf.multiply(labels, tf.log(predictions + epsilon)) - tf.multiply(
         (1 - labels), tf.log(1 - predictions + epsilon))
